% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{confint.desla}
\alias{confint.desla}
\title{Confidence intervals for desla objects}
\usage{
\method{confint}{desla}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{a \code{desla} object.}

\item{parm}{which parameters is the confidence interval needed for.}

\item{level}{confidence level(s).}

\item{...}{additional arguments (ignored).}
}
\description{
Confidence intervals for desla objects
}
\keyword{internal}
