% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{create_state_dummies}
\alias{create_state_dummies}
\title{Create State Dummies}
\usage{
create_state_dummies(x)
}
\arguments{
\item{x}{Contains the variables that define the states. Each column should either represent a categorical variable indicating the state of each observation, or each column should be a binary indicator for one particular state.}
}
\value{
A matrix where each column is a binary indicator for one state.
}
\description{
Creates state dummies for use in \code{\link{HDLP}}.
}
\details{
The function first checks if \code{x} is already in the correct output format by evaluating if each row sums up to one. If this is not the case, each column is treated as a categorical variable for which its unique entries define the states it can take. If \code{x} contains more than one column, interactions between the variables are created. Example, inputting two variables that can take two states each, results in a total of four possible states, and hence the output matrix contains four columns.
}
