% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{create_state_dummies_from_datamatrix}
\alias{create_state_dummies_from_datamatrix}
\title{Create State Dummies from Matrix}
\usage{
create_state_dummies_from_datamatrix(x)
}
\arguments{
\item{x}{Matrix or data frame where each column represents a state variable.}
}
\value{
A matrix where each column is a binary indicator for one state.
}
\description{
Creates state dummies from matrix-like objects
}
\keyword{internal}
