% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{create_state_dummies_from_vector}
\alias{create_state_dummies_from_vector}
\title{Create State Dummies from Vector}
\usage{
create_state_dummies_from_vector(x, varname = "StateVar")
}
\arguments{
\item{x}{Vector representing the state variable.}

\item{varname}{Name of the state variable.}
}
\value{
A matrix where each column is a binary indicator for one state.
}
\description{
Creates state dummies from vectors
}
\keyword{internal}
