% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{print.desla}
\alias{print.desla}
\title{Print desla output}
\usage{
\method{print}{desla}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  show_selected = 10,
  ...
)
}
\arguments{
\item{x}{a \code{desla} object.}

\item{digits}{digits.}

\item{signif.stars}{show stars of significance.}

\item{show_selected}{upper bound for which to show the names of selected variables in the
lasso regressions (default is 10)}

\item{...}{additional arguments (ignored).}
}
\description{
Print desla output
}
\keyword{internal}
