% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rwrap_build_gridsXy}
\alias{Rwrap_build_gridsXy}
\title{Generate inital and nodewise lambda grids}
\usage{
Rwrap_build_gridsXy(T, N, size, X, y, H)
}
\arguments{
\item{T}{number of rows in the X matrix}

\item{N}{number of columns in the X matrix}

\item{size}{number of lambdas in the grids}

\item{X}{regressor matrix}

\item{y}{dependent variable vector}

\item{H}{indexes of relevant regressors}
}
\description{
Generate inital and nodewise lambda grids
}
