% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rwrap_partial_desparsified_lasso_inference}
\alias{Rwrap_partial_desparsified_lasso_inference}
\title{Calculate the desparsified lasso}
\usage{
Rwrap_partial_desparsified_lasso_inference(
  X,
  y,
  H,
  init_partial,
  nw_partials,
  init_grid,
  nw_grids,
  init_selection_type,
  nw_selection_types,
  init_nonzero_limit,
  nw_nonzero_limits,
  init_opt_threshold,
  nw_opt_thresholds,
  init_opt_type,
  nw_opt_types,
  LRVtrunc,
  T_multiplier,
  alphas,
  R,
  q
)
}
\arguments{
\item{X}{regressor matrix}

\item{y}{dependent variable vector}

\item{H}{indexes of relevant regressors}

\item{init_partial}{(optional) boolean, true if you want the initial lasso to be partially penalized (false by default)}

\item{nw_partials}{(optional) boolean vector with the dimension of H, trues if you want the nodewise regressions to be partially penalized (all false by default)}

\item{init_grid}{(optional) vector, containing user specified initial grid}

\item{nw_grids}{(optional) matrix with number of rows the size of H, rows containing user specified grids for the nodewise regressions}

\item{init_selection_type}{(optional) integer, how should lambda be selected in the inital regression, 1=BIC, 2=AIC, 3=EBIC (1 by default)}

\item{nw_selection_types}{(optional) inteter vector with the dimension of H, how should lambda be selected in the nodewise regressions, 1=BIC, 2=AIC, 3=EBIC (all 1s by default)}

\item{init_nonzero_limit}{(optional) number controlling the maximum number of nonzeros that can be selected in the initial regression (0.5 by default, meaning no more than 0.5*T regressors can have nonzero estimates)}

\item{nw_nonzero_limits}{(optional) vector with the dimension of H, controlling the maximum number of nonzeros that can be selected in the nodewise regressions (0.5s by default)}

\item{init_opt_threshold}{(optional) optimization threshold for the coordinate descent algorithm in the inital regression (10^(-4) by default)}

\item{nw_opt_thresholds}{(optional) vector with the dimension of H, optimization thresholds for the coordinate descent algorithm in the nodewise lasso regression (10^(-4)s by default)}

\item{init_opt_type}{(optional) integer, which type of coordinate descent algorithm should be used in the inial regression, 1=naive, 2=covariance, 3=adaptive (3 by default)}

\item{nw_opt_types}{(optional)inteter vector with the dimension of H, which type of coordinate descent algorithm should be used in the nodewise regressions, 1=naive, 2=covariance, 3=adaptive (3s by default)}

\item{LRVtrunc}{(optional) parameter controlling the bandwidth Q_T used in the long run covariance matrix, Q_T=ceil(T_multiplier*T^LRVtrunc) (LRVtrunc=0.2 by default)}

\item{T_multiplier}{(optional) parameter controlling the bandwidth Q_T used in the long run covariance matrix, Q_T=ceil(T_multiplier*T^LRVtrunc) (Tmultiplier=2 by default)}

\item{alphas}{(optional) vector of significance levels (c(0.01,0.05,0.1) by default)}

\item{R}{(optional) matrix with number of columns the dimension of H, used to test the null hypothesis R*beta=q (identity matrix as default)}

\item{q}{(optional) vector of size same as the rows of H, used to test the null hypothesis R*beta=q (zeroes by default)}

\item{gridsize}{(optional) integer, how many different lambdas there should be in both inital and nodewise grids (100 by default)}
}
\description{
Calculate the desparsified lasso
}
