% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{summary.desla}
\alias{summary.desla}
\title{Summary of desla output}
\usage{
\method{summary}{desla}(object, ...)
}
\arguments{
\item{object}{a \code{desla} object.}

\item{...}{additional arguments (ignored).}
}
\description{
Summary of desla output
}
\keyword{internal}
