% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_functions.R
\name{coef.desla}
\alias{coef.desla}
\title{Extract coefficients}
\usage{
\method{coef}{desla}(object, ...)
}
\arguments{
\item{object}{a \code{desla} object.}

\item{...}{additional arguments (ignored).}

\item{scaled}{return the coefficients for the scaled variables (as used in the estimation)
if \code{TRUE} or the re-scaled coefficients for the original variables if \code{FALSE}
(default).}
}
\description{
Extract coefficients
}
\keyword{internal}
